# restfull/FEMS/swagger.yaml
openapi: 3.0.3
info:
  title: TZW LTD Fire Extinguisher Management API
  version: 1.0.0
  description: |
    Consolidated API documentation for the FEMS backend services behind the API gateway.

    Covered services:
    - Authentication and user administration
    - Facility and customer management
    - Fire extinguisher registration and tracking
    - Inspection scheduling and recording
    - Maintenance logging
    - Notifications and escalations
    - Reporting and audit access
servers:
  - url: http://localhost:5000/api
    description: Local API gateway
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit10Param:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    SearchParam:
      name: search
      in: query
      schema:
        type: string
        default: ''
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Extinguisher:
      type: object
      properties:
        id:
          type: string
          format: uuid
        extinguisherCode:
          type: string
        serialNumber:
          type: string
        type:
          type: string
          enum: [Water, CO2, Foam, Dry Chemical]
        size:
          type: string
          enum: ['2.5 lb', '5 lb', '9 lb', '12 lb']
        capacityLiters:
          type: number
          nullable: true
        manufactureDate:
          type: string
          format: date
          nullable: true
        purchaseDate:
          type: string
          format: date
          nullable: true
        installationDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        lastInspectionDate:
          type: string
          format: date
          nullable: true
        nextInspectionDate:
          type: string
          format: date
          nullable: true
        location:
          type: string
        customerId:
          type: string
          format: uuid
        customerName:
          type: string
        customerOrg:
          type: string
          nullable: true
        customerEmail:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, expired, serviced, decommissioned, pending_inspection]
        complianceStatus:
          type: string
          enum: [compliant, warning, critical, non_compliant]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
    ExtinguisherCreateRequest:
      type: object
      required: [serialNumber, type, size, location, installationDate, expiryDate, customerId]
      properties:
        serialNumber:
          type: string
        type:
          type: string
          enum: [Water, CO2, Foam, Dry Chemical]
        size:
          type: string
          enum: ['2.5 lb', '5 lb', '9 lb', '12 lb']
        location:
          type: string
        installationDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, expired, serviced, decommissioned, pending_inspection]
          default: active
        customerId:
          type: string
          format: uuid
        manufactureDate:
          type: string
          format: date
          nullable: true
        purchaseDate:
          type: string
          format: date
          nullable: true
        capacityLiters:
          type: number
          nullable: true
        lastInspectionDate:
          type: string
          format: date
          nullable: true
        nextInspectionDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
    ExtinguisherUpdateRequest:
      type: object
      properties:
        serialNumber:
          type: string
        type:
          type: string
          enum: [Water, CO2, Foam, Dry Chemical]
        size:
          type: string
          enum: ['2.5 lb', '5 lb', '9 lb', '12 lb']
        location:
          type: string
        installationDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, expired, serviced, decommissioned, pending_inspection]
        manufactureDate:
          type: string
          format: date
          nullable: true
        purchaseDate:
          type: string
          format: date
          nullable: true
        capacityLiters:
          type: number
          nullable: true
        lastInspectionDate:
          type: string
          format: date
          nullable: true
        nextInspectionDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
    NotificationStatus:
      type: object
      properties:
        emailSent:
          type: boolean
        emailRecipient:
          type: string
          nullable: true
        emailError:
          type: string
          nullable: true
...
security:
  - bearerAuth: []
paths:
...
  /extinguishers:
    post:
      tags: [Extinguishers]
      summary: Register a fire extinguisher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtinguisherCreateRequest'
      responses:
        '201':
          description: Extinguisher registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  data:
                    $ref: '#/components/schemas/Extinguisher'
                  notification:
                    $ref: '#/components/schemas/NotificationStatus'
    get:
      tags: [Extinguishers]
      summary: List extinguishers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/Limit10Param'
        - $ref: '#/components/parameters/SearchParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, expired, serviced, decommissioned, pending_inspection]
        - name: type
          in: query
          schema:
            type: string
            enum: [Water, CO2, Foam, Dry Chemical]
        - name: customerId
          in: query
          schema:
            type: string
            format: uuid
        - name: compliance
          in: query
          schema:
            type: string
            enum: [compliant, warning, critical, non_compliant]
      responses:
        '200':
          description: Extinguisher list
...
  /extinguishers/{id}:
    get:
      tags: [Extinguishers]
      summary: Get extinguisher details
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Extinguisher details
    put:
      tags: [Extinguishers]
      summary: Update extinguisher details
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtinguisherUpdateRequest'
      responses:
        '200':
          description: Extinguisher updated
    delete:
      tags: [Extinguishers]
      summary: Delete an extinguisher
      description: Admin only
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Extinguisher removed
...